+++ src/components/Dropdown/Dropdown.tsx
@@ -282,7 +282,6 @@ export const Dropdown: FC<DropdownProps> = React.memo(
       };
 
       const handleFloatingKeyDown = (event: React.KeyboardEvent): void => {
-        event.stopPropagation();
         if (event?.key === eventKeys.ESCAPE) {
           toggle(false)(event);
         }
@@ -321,9 +320,12 @@ export const Dropdown: FC<DropdownProps> = React.memo(
         }
         // If there's an ariaRef, apply the a11y attributes to it, rather than the immediate child.
         if (ariaRef?.current) {
-          ariaRef.current.setAttribute('aria-controls', dropdownId);
           ariaRef.current.setAttribute('aria-expanded', `${mergedVisible}`);
           ariaRef.current.setAttribute('aria-haspopup', 'true');
+
+          if (!ariaRef.current.hasAttribute('aria-controls')) {
+            ariaRef.current.setAttribute('aria-controls', dropdownId);
+          }
 
           if (!ariaRef.current.hasAttribute('role')) {
             ariaRef.current.setAttribute('role', 'button');
@@ -420,7 +422,9 @@ export const Dropdown: FC<DropdownProps> = React.memo(
           {getReference()}
           <ConditionalWrapper
             condition={portal}
-            wrapper={(children) => <FloatingPortal>{children}</FloatingPortal>}
+            wrapper={(children) => (
+              <FloatingPortal preserveTabOrder>{children}</FloatingPortal>
+            )}
           >
             {getDropdown()}
           </ConditionalWrapper>