+++ CHANGELOG.md
@@ -2,6 +2,41 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [2.52.18](https://github.com/EightfoldAI/octuple/compare/v2.52.17...v2.52.18) (2024-11-13)
+
+### Bug Fixes
+
+- dropdown items are not moving to the next tab index due to stopProgation ([#911](https://github.com/EightfoldAI/octuple/issues/911)) ([d571817](https://github.com/EightfoldAI/octuple/commits/d571817c0b10e409ecdc18fd0913150cfc63c663))
+
+### [2.52.17](https://github.com/EightfoldAI/octuple/compare/v2.52.15-1...v2.52.17) (2024-11-12)
+
+### Bug Fixes
+
+- auto close date picker after selection ([#907](https://github.com/EightfoldAI/octuple/issues/907)) ([5c8c0ce](https://github.com/EightfoldAI/octuple/commits/5c8c0cefe989e9cc6c78ff8f8cf5a59ae7092843))
+- dropdown auto click the element when TAB operation is invoked ([#909](https://github.com/EightfoldAI/octuple/issues/909)) ([a86ce4d](https://github.com/EightfoldAI/octuple/commits/a86ce4dd8c8341af7588ff862294c2ec3f741261))
+
+### [2.52.16](https://github.com/EightfoldAI/octuple/compare/v2.52.15-1...v2.52.16) (2024-11-11)
+
+### Bug Fixes
+
+- auto close date picker after selection ([#907](https://github.com/EightfoldAI/octuple/issues/907)) ([5c8c0ce](https://github.com/EightfoldAI/octuple/commits/5c8c0cefe989e9cc6c78ff8f8cf5a59ae7092843))
+
+### [2.52.15](https://github.com/EightfoldAI/octuple/compare/v2.52.15-1...v2.52.15) (2024-11-11)
+
+### [2.52.15-1](https://github.com/EightfoldAI/octuple/compare/v2.52.14...v2.52.15-1) (2024-11-11)
+
+### Features
+
+- change version in package.json ([#905](https://github.com/EightfoldAI/octuple/issues/905)) ([0e1730e](https://github.com/EightfoldAI/octuple/commits/0e1730eaf9b6fe076110644ee8a158a04ba20867))
+- remove focus visible ([#906](https://github.com/EightfoldAI/octuple/issues/906)) ([dd6eb34](https://github.com/EightfoldAI/octuple/commits/dd6eb3493afc703d555787ff193cc74a883ff598))
+
+### [2.52.15-0](https://github.com/EightfoldAI/octuple/compare/v2.52.14...v2.52.15-0) (2024-11-11)
+
+### Features
+
+- change version in package.json ([#905](https://github.com/EightfoldAI/octuple/issues/905)) ([0e1730e](https://github.com/EightfoldAI/octuple/commits/0e1730eaf9b6fe076110644ee8a158a04ba20867))
+- remove focus visible ([#906](https://github.com/EightfoldAI/octuple/issues/906)) ([dd6eb34](https://github.com/EightfoldAI/octuple/commits/dd6eb3493afc703d555787ff193cc74a883ff598))
+
 ## [2.52.0-rollup-rc.0](https://github.com/EightfoldAI/octuple/compare/v2.51.3...v2.52.0-rollup-rc.0) (2024-06-24)
 
 ### Features